remove LogEntry on RefreshOrderView
RefreshOrderView contain an hardcoded User.id to log an action. But this user may not exist and even if it's the case, we have poor chance it's really the one doing the action. I removed the LogEntry and open the discussion in a PR to see what we should do.